@@ -18,12 +18,11 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.19.4 <=3.3.8"
 
 dependencies:
   flutter:
     sdk: flutter
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -32,16 +31,19 @@ dependencies:
   provider: ^6.0.3
   dartz: ^0.10.0-nullsafety.2
   equatable: ^2.0.2
-  google_fonts: ^3.0.1
+  google_fonts: ^4.0.3
   flutter_rating_bar: ^4.0.0
   cached_network_image: ^3.0.0
   get_it: ^7.1.3
   sqflite: ^2.0.0+3
   path_provider: ^2.0.2
+  dynamic_color: ^1.6.2
 
 dev_dependencies:
   mockito: ^5.0.8
   build_runner: ^2.0.4
+  integration_test:
+    sdk: flutter
   flutter_test:
     sdk: flutter
 
