@@ -1,5 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:ditonton/domain/usecases/get_watchlist_movies.dart';
+import 'package:ditonton/domain/usecases/get_movie_watchlist.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
@@ -7,12 +7,12 @@ import '../../dummy_data/dummy_objects.dart';
 import '../../helpers/test_helper.mocks.dart';
 
 void main() {
-  late GetWatchlistMovies usecase;
+  late GetMovieWatchlist usecase;
   late MockMovieRepository mockMovieRepository;
 
   setUp(() {
     mockMovieRepository = MockMovieRepository();
-    usecase = GetWatchlistMovies(mockMovieRepository);
+    usecase = GetMovieWatchlist(mockMovieRepository);
   });
 
   test('should get list of movies from the repository', () async {
@@ -20,7 +20,7 @@ void main() {
     when(mockMovieRepository.getWatchlistMovies())
         .thenAnswer((_) async => Right(testMovieList));
     // act
-    final result = await usecase.execute();
+    final result = await usecase();
     // assert
     expect(result, Right(testMovieList));
   });
