@@ -1,16 +1,16 @@
-import 'package:ditonton/domain/usecases/get_watchlist_status.dart';
+import 'package:ditonton/domain/usecases/get_movie_watchlist_status.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../helpers/test_helper.mocks.dart';
 
 void main() {
-  late GetWatchListStatus usecase;
+  late GetMovieWatchListStatus usecase;
   late MockMovieRepository mockMovieRepository;
 
   setUp(() {
     mockMovieRepository = MockMovieRepository();
-    usecase = GetWatchListStatus(mockMovieRepository);
+    usecase = GetMovieWatchListStatus(mockMovieRepository);
   });
 
   test('should get watchlist status from repository', () async {
@@ -18,7 +18,7 @@ void main() {
     when(mockMovieRepository.isAddedToWatchlist(1))
         .thenAnswer((_) async => true);
     // act
-    final result = await usecase.execute(1);
+    final result = await usecase(1);
     // assert
     expect(result, true);
   });
