@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
 
 const String BASE_IMAGE_URL = 'https://image.tmdb.org/t/p/w500';
 
@@ -10,24 +9,6 @@ const Color kPrussianBlue = Color(0xFF003566);
 const Color kMikadoYellow = Color(0xFFffc300);
 const Color kDavysGrey = Color(0xFF4B5358);
 const Color kGrey = Color(0xFF303030);
-
-// text style
-final TextStyle kHeading5 =
-    GoogleFonts.poppins(fontSize: 23, fontWeight: FontWeight.w400);
-final TextStyle kHeading6 = GoogleFonts.poppins(
-    fontSize: 19, fontWeight: FontWeight.w500, letterSpacing: 0.15);
-final TextStyle kSubtitle = GoogleFonts.poppins(
-    fontSize: 15, fontWeight: FontWeight.w400, letterSpacing: 0.15);
-final TextStyle kBodyText = GoogleFonts.poppins(
-    fontSize: 13, fontWeight: FontWeight.w400, letterSpacing: 0.25);
-
-// text theme
-final kTextTheme = TextTheme(
-  headline5: kHeading5,
-  headline6: kHeading6,
-  subtitle1: kSubtitle,
-  bodyText2: kBodyText,
-);
 
 const kColorScheme = ColorScheme(
   primary: kMikadoYellow,
@@ -44,3 +25,14 @@ const kColorScheme = ColorScheme(
   onError: Colors.white,
   brightness: Brightness.dark,
 );
+
+/// Messages
+const kWatchlistAddSuccessMessage = 'Added to Watchlist';
+const kWatchlistRemoveSuccessMessage = 'Removed from Watchlist';
+
+/// Widget keys
+const kWatchlistButtonKey = Key('watchlist_button');
+const kShimmerKey = Key('shimmer');
+const kSimpleEntertainmentCardKey = Key('simple_entertainment_card');
+const kSearchButtonKey = Key('search_button');
+const kTexFieldKey = Key('text_field');
