@@ -1,19 +1,14 @@
+import 'package:ditonton/data/models/watchlist_table.dart';
 import 'package:ditonton/domain/entities/movie.dart';
 import 'package:ditonton/domain/entities/movie_detail.dart';
-import 'package:equatable/equatable.dart';
 
-class MovieTable extends Equatable {
-  final int id;
-  final String? title;
-  final String? posterPath;
-  final String? overview;
-
-  MovieTable({
-    required this.id,
-    required this.title,
-    required this.posterPath,
-    required this.overview,
-  });
+class MovieTable extends WatchlistTable {
+  const MovieTable({
+    required super.id,
+    required super.title,
+    required super.posterPath,
+    required super.overview,
+  }) : super(type: 'movie');
 
   factory MovieTable.fromEntity(MovieDetail movie) => MovieTable(
         id: movie.id,
@@ -29,21 +24,10 @@ class MovieTable extends Equatable {
         overview: map['overview'],
       );
 
-  Map<String, dynamic> toJson() => {
-        'id': id,
-        'title': title,
-        'posterPath': posterPath,
-        'overview': overview,
-      };
-
-  Movie toEntity() => Movie.watchlist(
+  Movie toEntity() => Movie(
         id: id,
         overview: overview,
         posterPath: posterPath,
         title: title,
       );
-
-  @override
-  // TODO: implement props
-  List<Object?> get props => [id, title, posterPath, overview];
 }
