@@ -1,14 +1,14 @@
 import 'package:dartz/dartz.dart';
+import 'package:ditonton/common/failure.dart';
 import 'package:ditonton/domain/entities/movie_detail.dart';
 import 'package:ditonton/domain/repositories/movie_repository.dart';
-import 'package:ditonton/common/failure.dart';
 
 class GetMovieDetail {
   final MovieRepository repository;
 
   GetMovieDetail(this.repository);
 
-  Future<Either<Failure, MovieDetail>> execute(int id) {
+  Future<Either<Failure, MovieDetail>> call(int id) {
     return repository.getMovieDetail(id);
   }
 }
