@@ -1,14 +1,14 @@
 import 'package:dartz/dartz.dart';
+import 'package:ditonton/common/failure.dart';
 import 'package:ditonton/domain/entities/movie.dart';
 import 'package:ditonton/domain/repositories/movie_repository.dart';
-import 'package:ditonton/common/failure.dart';
 
 class GetMovieRecommendations {
   final MovieRepository repository;
 
   GetMovieRecommendations(this.repository);
 
-  Future<Either<Failure, List<Movie>>> execute(id) {
+  Future<Either<Failure, List<Movie>>> call(id, [int page = 1]) {
     return repository.getMovieRecommendations(id);
   }
 }
