@@ -1,14 +1,14 @@
 import 'package:dartz/dartz.dart';
+import 'package:ditonton/common/failure.dart';
 import 'package:ditonton/domain/entities/movie.dart';
 import 'package:ditonton/domain/repositories/movie_repository.dart';
-import 'package:ditonton/common/failure.dart';
 
 class GetNowPlayingMovies {
   final MovieRepository repository;
 
   GetNowPlayingMovies(this.repository);
 
-  Future<Either<Failure, List<Movie>>> execute() {
+  Future<Either<Failure, List<Movie>>> call([int page = 1]) {
     return repository.getNowPlayingMovies();
   }
 }
