@@ -1,37 +1,48 @@
-import 'package:ditonton/domain/entities/movie.dart';
-import 'package:ditonton/domain/usecases/get_now_playing_movies.dart';
 import 'package:ditonton/common/state_enum.dart';
+import 'package:ditonton/domain/entities/movie.dart';
+import 'package:ditonton/domain/entities/watchlist.dart';
+import 'package:ditonton/domain/usecases/get_now_playing_movies.dart';
 import 'package:ditonton/domain/usecases/get_popular_movies.dart';
 import 'package:ditonton/domain/usecases/get_top_rated_movies.dart';
 import 'package:flutter/material.dart';
 
 class MovieListNotifier extends ChangeNotifier {
   var _nowPlayingMovies = <Movie>[];
-  List<Movie> get nowPlayingMovies => _nowPlayingMovies;
+
+  List<Watchlist> get nowPlayingMovies =>
+      _nowPlayingMovies.map((data) => data.toWatchlist()).toList();
 
   RequestState _nowPlayingState = RequestState.Empty;
+
   RequestState get nowPlayingState => _nowPlayingState;
 
   var _popularMovies = <Movie>[];
-  List<Movie> get popularMovies => _popularMovies;
+
+  List<Watchlist> get popularMovies =>
+      _popularMovies.map((data) => data.toWatchlist()).toList();
 
   RequestState _popularMoviesState = RequestState.Empty;
+
   RequestState get popularMoviesState => _popularMoviesState;
 
   var _topRatedMovies = <Movie>[];
-  List<Movie> get topRatedMovies => _topRatedMovies;
+
+  List<Watchlist> get topRatedMovies =>
+      _topRatedMovies.map((data) => data.toWatchlist()).toList();
 
   RequestState _topRatedMoviesState = RequestState.Empty;
+
   RequestState get topRatedMoviesState => _topRatedMoviesState;
 
   String _message = '';
+
   String get message => _message;
 
-  MovieListNotifier({
-    required this.getNowPlayingMovies,
-    required this.getPopularMovies,
-    required this.getTopRatedMovies,
-  });
+  MovieListNotifier(
+    this.getNowPlayingMovies,
+    this.getPopularMovies,
+    this.getTopRatedMovies,
+  );
 
   final GetNowPlayingMovies getNowPlayingMovies;
   final GetPopularMovies getPopularMovies;
@@ -41,7 +52,7 @@ class MovieListNotifier extends ChangeNotifier {
     _nowPlayingState = RequestState.Loading;
     notifyListeners();
 
-    final result = await getNowPlayingMovies.execute();
+    final result = await getNowPlayingMovies();
     result.fold(
       (failure) {
         _nowPlayingState = RequestState.Error;
@@ -60,7 +71,7 @@ class MovieListNotifier extends ChangeNotifier {
     _popularMoviesState = RequestState.Loading;
     notifyListeners();
 
-    final result = await getPopularMovies.execute();
+    final result = await getPopularMovies();
     result.fold(
       (failure) {
         _popularMoviesState = RequestState.Error;
@@ -79,7 +90,7 @@ class MovieListNotifier extends ChangeNotifier {
     _topRatedMoviesState = RequestState.Loading;
     notifyListeners();
 
-    final result = await getTopRatedMovies.execute();
+    final result = await getTopRatedMovies();
     result.fold(
       (failure) {
         _topRatedMoviesState = RequestState.Error;
