@@ -22,7 +22,7 @@ void main() {
     when(mockMovieRepository.getMovieDetail(tId))
         .thenAnswer((_) async => Right(testMovieDetail));
     // act
-    final result = await usecase.execute(tId);
+    final result = await usecase(tId);
     // assert
     expect(result, Right(testMovieDetail));
   });
