@@ -1,14 +1,20 @@
-import 'package:ditonton/data/datasources/db/database_helper.dart';
+import 'package:ditonton/data/datasources/database_helper.dart';
 import 'package:ditonton/data/datasources/movie_local_data_source.dart';
 import 'package:ditonton/data/datasources/movie_remote_data_source.dart';
+import 'package:ditonton/data/datasources/tv_show_local_data_source.dart';
+import 'package:ditonton/data/datasources/tv_show_remote_data_source.dart';
 import 'package:ditonton/domain/repositories/movie_repository.dart';
-import 'package:mockito/annotations.dart';
+import 'package:ditonton/domain/repositories/tv_repository.dart';
 import 'package:http/http.dart' as http;
+import 'package:mockito/annotations.dart';
 
 @GenerateMocks([
   MovieRepository,
   MovieRemoteDataSource,
   MovieLocalDataSource,
+  TvShowRepository,
+  TvShowRemoteDataSource,
+  TvShowLocalDataSource,
   DatabaseHelper,
 ], customMocks: [
   MockSpec<http.Client>(as: #MockHttpClient)
