@@ -1,44 +1,102 @@
-// Mocks generated by Mockito 5.0.8 from annotations
+// Mocks generated by Mockito 5.3.2 from annotations
 // in ditonton/test/helpers/test_helper.dart.
 // Do not manually edit this file.
 
+// ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:async' as _i7;
-import 'dart:convert' as _i18;
-import 'dart:typed_data' as _i19;
+import 'dart:convert' as _i25;
+import 'dart:typed_data' as _i26;
 
 import 'package:dartz/dartz.dart' as _i2;
 import 'package:ditonton/common/failure.dart' as _i8;
-import 'package:ditonton/data/datasources/db/database_helper.dart' as _i15;
+import 'package:ditonton/data/datasources/database_helper.dart' as _i22;
 import 'package:ditonton/data/datasources/movie_local_data_source.dart' as _i13;
 import 'package:ditonton/data/datasources/movie_remote_data_source.dart'
     as _i11;
+import 'package:ditonton/data/datasources/tv_show_local_data_source.dart'
+    as _i20;
+import 'package:ditonton/data/datasources/tv_show_remote_data_source.dart'
+    as _i18;
 import 'package:ditonton/data/models/movie_detail_model.dart' as _i3;
 import 'package:ditonton/data/models/movie_model.dart' as _i12;
 import 'package:ditonton/data/models/movie_table.dart' as _i14;
+import 'package:ditonton/data/models/tv_show_detail_model.dart' as _i4;
+import 'package:ditonton/data/models/tv_show_model.dart' as _i19;
+import 'package:ditonton/data/models/tv_show_table.dart' as _i21;
+import 'package:ditonton/data/models/watchlist_table.dart' as _i24;
 import 'package:ditonton/domain/entities/movie.dart' as _i9;
 import 'package:ditonton/domain/entities/movie_detail.dart' as _i10;
+import 'package:ditonton/domain/entities/tv_show.dart' as _i16;
+import 'package:ditonton/domain/entities/tv_show_detail.dart' as _i17;
 import 'package:ditonton/domain/repositories/movie_repository.dart' as _i6;
-import 'package:http/src/base_request.dart' as _i20;
-import 'package:http/src/client.dart' as _i17;
-import 'package:http/src/response.dart' as _i4;
-import 'package:http/src/streamed_response.dart' as _i5;
+import 'package:ditonton/domain/repositories/tv_repository.dart' as _i15;
+import 'package:http/http.dart' as _i5;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:sqflite_common/sqlite_api.dart' as _i16;
+import 'package:sqflite/sqflite.dart' as _i23;
 
+// ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
 // ignore_for_file: comment_references
+// ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: prefer_const_constructors
 // ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
 
-class _FakeEither<L, R> extends _i1.Fake implements _i2.Either<L, R> {}
+class _FakeEither_0<L, R> extends _i1.SmartFake implements _i2.Either<L, R> {
+  _FakeEither_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeMovieDetailResponse extends _i1.Fake
-    implements _i3.MovieDetailResponse {}
+class _FakeMovieDetailResponse_1 extends _i1.SmartFake
+    implements _i3.MovieDetailResponse {
+  _FakeMovieDetailResponse_1(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeResponse extends _i1.Fake implements _i4.Response {}
+class _FakeTvShowDetailModel_2 extends _i1.SmartFake
+    implements _i4.TvShowDetailModel {
+  _FakeTvShowDetailModel_2(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
-class _FakeStreamedResponse extends _i1.Fake implements _i5.StreamedResponse {}
+class _FakeResponse_3 extends _i1.SmartFake implements _i5.Response {
+  _FakeResponse_3(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
+class _FakeStreamedResponse_4 extends _i1.SmartFake
+    implements _i5.StreamedResponse {
+  _FakeStreamedResponse_4(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
 
 /// A class which mocks [MovieRepository].
 ///
@@ -50,67 +108,162 @@ class MockMovieRepository extends _i1.Mock implements _i6.MovieRepository {
 
   @override
   _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>> getNowPlayingMovies() =>
-      (super.noSuchMethod(Invocation.method(#getNowPlayingMovies, []),
-          returnValue: Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
-              _FakeEither<_i8.Failure, List<_i9.Movie>>())) as _i7
-          .Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getNowPlayingMovies,
+          [],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
+            _FakeEither_0<_i8.Failure, List<_i9.Movie>>(
+          this,
+          Invocation.method(
+            #getNowPlayingMovies,
+            [],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>> getPopularMovies() =>
-      (super.noSuchMethod(Invocation.method(#getPopularMovies, []),
-          returnValue: Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
-              _FakeEither<_i8.Failure, List<_i9.Movie>>())) as _i7
-          .Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getPopularMovies,
+          [],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
+            _FakeEither_0<_i8.Failure, List<_i9.Movie>>(
+          this,
+          Invocation.method(
+            #getPopularMovies,
+            [],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>> getTopRatedMovies() =>
-      (super.noSuchMethod(Invocation.method(#getTopRatedMovies, []),
-          returnValue: Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
-              _FakeEither<_i8.Failure, List<_i9.Movie>>())) as _i7
-          .Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTopRatedMovies,
+          [],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
+            _FakeEither_0<_i8.Failure, List<_i9.Movie>>(
+          this,
+          Invocation.method(
+            #getTopRatedMovies,
+            [],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, _i10.MovieDetail>> getMovieDetail(
           int? id) =>
-      (super.noSuchMethod(Invocation.method(#getMovieDetail, [id]),
-          returnValue: Future<_i2.Either<_i8.Failure, _i10.MovieDetail>>.value(
-              _FakeEither<_i8.Failure, _i10.MovieDetail>())) as _i7
-          .Future<_i2.Either<_i8.Failure, _i10.MovieDetail>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMovieDetail,
+          [id],
+        ),
+        returnValue:
+            _i7.Future<_i2.Either<_i8.Failure, _i10.MovieDetail>>.value(
+                _FakeEither_0<_i8.Failure, _i10.MovieDetail>(
+          this,
+          Invocation.method(
+            #getMovieDetail,
+            [id],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, _i10.MovieDetail>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>> getMovieRecommendations(
           int? id) =>
-      (super.noSuchMethod(Invocation.method(#getMovieRecommendations, [id]),
-          returnValue: Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
-              _FakeEither<_i8.Failure, List<_i9.Movie>>())) as _i7
-          .Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMovieRecommendations,
+          [id],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
+            _FakeEither_0<_i8.Failure, List<_i9.Movie>>(
+          this,
+          Invocation.method(
+            #getMovieRecommendations,
+            [id],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>> searchMovies(
           String? query) =>
-      (super.noSuchMethod(Invocation.method(#searchMovies, [query]),
-          returnValue: Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
-              _FakeEither<_i8.Failure, List<_i9.Movie>>())) as _i7
-          .Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #searchMovies,
+          [query],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
+            _FakeEither_0<_i8.Failure, List<_i9.Movie>>(
+          this,
+          Invocation.method(
+            #searchMovies,
+            [query],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, String>> saveWatchlist(
           _i10.MovieDetail? movie) =>
-      (super.noSuchMethod(Invocation.method(#saveWatchlist, [movie]),
-              returnValue: Future<_i2.Either<_i8.Failure, String>>.value(
-                  _FakeEither<_i8.Failure, String>()))
-          as _i7.Future<_i2.Either<_i8.Failure, String>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #saveWatchlist,
+          [movie],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, String>>.value(
+            _FakeEither_0<_i8.Failure, String>(
+          this,
+          Invocation.method(
+            #saveWatchlist,
+            [movie],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, String>>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, String>> removeWatchlist(
           _i10.MovieDetail? movie) =>
-      (super.noSuchMethod(Invocation.method(#removeWatchlist, [movie]),
-              returnValue: Future<_i2.Either<_i8.Failure, String>>.value(
-                  _FakeEither<_i8.Failure, String>()))
-          as _i7.Future<_i2.Either<_i8.Failure, String>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #removeWatchlist,
+          [movie],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, String>>.value(
+            _FakeEither_0<_i8.Failure, String>(
+          this,
+          Invocation.method(
+            #removeWatchlist,
+            [movie],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, String>>);
   @override
-  _i7.Future<bool> isAddedToWatchlist(int? id) =>
-      (super.noSuchMethod(Invocation.method(#isAddedToWatchlist, [id]),
-          returnValue: Future<bool>.value(false)) as _i7.Future<bool>);
+  _i7.Future<bool> isAddedToWatchlist(int? id) => (super.noSuchMethod(
+        Invocation.method(
+          #isAddedToWatchlist,
+          [id],
+        ),
+        returnValue: _i7.Future<bool>.value(false),
+      ) as _i7.Future<bool>);
   @override
   _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>> getWatchlistMovies() =>
-      (super.noSuchMethod(Invocation.method(#getWatchlistMovies, []),
-          returnValue: Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
-              _FakeEither<_i8.Failure, List<_i9.Movie>>())) as _i7
-          .Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getWatchlistMovies,
+          [],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>.value(
+            _FakeEither_0<_i8.Failure, List<_i9.Movie>>(
+          this,
+          Invocation.method(
+            #getWatchlistMovies,
+            [],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i9.Movie>>>);
 }
 
 /// A class which mocks [MovieRemoteDataSource].
@@ -124,38 +277,68 @@ class MockMovieRemoteDataSource extends _i1.Mock
 
   @override
   _i7.Future<List<_i12.MovieModel>> getNowPlayingMovies() =>
-      (super.noSuchMethod(Invocation.method(#getNowPlayingMovies, []),
-              returnValue:
-                  Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]))
-          as _i7.Future<List<_i12.MovieModel>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getNowPlayingMovies,
+          [],
+        ),
+        returnValue:
+            _i7.Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]),
+      ) as _i7.Future<List<_i12.MovieModel>>);
   @override
   _i7.Future<List<_i12.MovieModel>> getPopularMovies() => (super.noSuchMethod(
-          Invocation.method(#getPopularMovies, []),
-          returnValue: Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]))
-      as _i7.Future<List<_i12.MovieModel>>);
+        Invocation.method(
+          #getPopularMovies,
+          [],
+        ),
+        returnValue:
+            _i7.Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]),
+      ) as _i7.Future<List<_i12.MovieModel>>);
   @override
   _i7.Future<List<_i12.MovieModel>> getTopRatedMovies() => (super.noSuchMethod(
-          Invocation.method(#getTopRatedMovies, []),
-          returnValue: Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]))
-      as _i7.Future<List<_i12.MovieModel>>);
+        Invocation.method(
+          #getTopRatedMovies,
+          [],
+        ),
+        returnValue:
+            _i7.Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]),
+      ) as _i7.Future<List<_i12.MovieModel>>);
   @override
   _i7.Future<_i3.MovieDetailResponse> getMovieDetail(int? id) =>
-      (super.noSuchMethod(Invocation.method(#getMovieDetail, [id]),
-              returnValue: Future<_i3.MovieDetailResponse>.value(
-                  _FakeMovieDetailResponse()))
-          as _i7.Future<_i3.MovieDetailResponse>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMovieDetail,
+          [id],
+        ),
+        returnValue: _i7.Future<_i3.MovieDetailResponse>.value(
+            _FakeMovieDetailResponse_1(
+          this,
+          Invocation.method(
+            #getMovieDetail,
+            [id],
+          ),
+        )),
+      ) as _i7.Future<_i3.MovieDetailResponse>);
   @override
   _i7.Future<List<_i12.MovieModel>> getMovieRecommendations(int? id) =>
-      (super.noSuchMethod(Invocation.method(#getMovieRecommendations, [id]),
-              returnValue:
-                  Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]))
-          as _i7.Future<List<_i12.MovieModel>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #getMovieRecommendations,
+          [id],
+        ),
+        returnValue:
+            _i7.Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]),
+      ) as _i7.Future<List<_i12.MovieModel>>);
   @override
   _i7.Future<List<_i12.MovieModel>> searchMovies(String? query) =>
-      (super.noSuchMethod(Invocation.method(#searchMovies, [query]),
-              returnValue:
-                  Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]))
-          as _i7.Future<List<_i12.MovieModel>>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #searchMovies,
+          [query],
+        ),
+        returnValue:
+            _i7.Future<List<_i12.MovieModel>>.value(<_i12.MovieModel>[]),
+      ) as _i7.Future<List<_i12.MovieModel>>);
 }
 
 /// A class which mocks [MovieLocalDataSource].
@@ -169,134 +352,645 @@ class MockMovieLocalDataSource extends _i1.Mock
 
   @override
   _i7.Future<String> insertWatchlist(_i14.MovieTable? movie) =>
-      (super.noSuchMethod(Invocation.method(#insertWatchlist, [movie]),
-          returnValue: Future<String>.value('')) as _i7.Future<String>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #insertWatchlist,
+          [movie],
+        ),
+        returnValue: _i7.Future<String>.value(''),
+      ) as _i7.Future<String>);
   @override
   _i7.Future<String> removeWatchlist(_i14.MovieTable? movie) =>
-      (super.noSuchMethod(Invocation.method(#removeWatchlist, [movie]),
-          returnValue: Future<String>.value('')) as _i7.Future<String>);
+      (super.noSuchMethod(
+        Invocation.method(
+          #removeWatchlist,
+          [movie],
+        ),
+        returnValue: _i7.Future<String>.value(''),
+      ) as _i7.Future<String>);
   @override
-  _i7.Future<_i14.MovieTable?> getMovieById(int? id) =>
-      (super.noSuchMethod(Invocation.method(#getMovieById, [id]),
-              returnValue: Future<_i14.MovieTable?>.value())
-          as _i7.Future<_i14.MovieTable?>);
+  _i7.Future<_i14.MovieTable?> getMovieById(int? id) => (super.noSuchMethod(
+        Invocation.method(
+          #getMovieById,
+          [id],
+        ),
+        returnValue: _i7.Future<_i14.MovieTable?>.value(),
+      ) as _i7.Future<_i14.MovieTable?>);
   @override
   _i7.Future<List<_i14.MovieTable>> getWatchlistMovies() => (super.noSuchMethod(
-          Invocation.method(#getWatchlistMovies, []),
-          returnValue: Future<List<_i14.MovieTable>>.value(<_i14.MovieTable>[]))
-      as _i7.Future<List<_i14.MovieTable>>);
+        Invocation.method(
+          #getWatchlistMovies,
+          [],
+        ),
+        returnValue:
+            _i7.Future<List<_i14.MovieTable>>.value(<_i14.MovieTable>[]),
+      ) as _i7.Future<List<_i14.MovieTable>>);
+}
+
+/// A class which mocks [TvShowRepository].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockTvShowRepository extends _i1.Mock implements _i15.TvShowRepository {
+  MockTvShowRepository() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>> getNowAiringTvShows(
+          [int? page = 1]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getNowAiringTvShows,
+          [page],
+        ),
+        returnValue:
+            _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>.value(
+                _FakeEither_0<_i8.Failure, List<_i16.TvShow>>(
+          this,
+          Invocation.method(
+            #getNowAiringTvShows,
+            [page],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>> getPopularTvShows(
+          [int? page = 1]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getPopularTvShows,
+          [page],
+        ),
+        returnValue:
+            _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>.value(
+                _FakeEither_0<_i8.Failure, List<_i16.TvShow>>(
+          this,
+          Invocation.method(
+            #getPopularTvShows,
+            [page],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>> getTopRatedTvShows(
+          [int? page = 1]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTopRatedTvShows,
+          [page],
+        ),
+        returnValue:
+            _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>.value(
+                _FakeEither_0<_i8.Failure, List<_i16.TvShow>>(
+          this,
+          Invocation.method(
+            #getTopRatedTvShows,
+            [page],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, _i17.TvShowDetail>> getTvShowDetail(
+          int? id) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTvShowDetail,
+          [id],
+        ),
+        returnValue:
+            _i7.Future<_i2.Either<_i8.Failure, _i17.TvShowDetail>>.value(
+                _FakeEither_0<_i8.Failure, _i17.TvShowDetail>(
+          this,
+          Invocation.method(
+            #getTvShowDetail,
+            [id],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, _i17.TvShowDetail>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>
+      getTvShowRecommendations(
+    int? id, [
+    int? page = 1,
+  ]) =>
+          (super.noSuchMethod(
+            Invocation.method(
+              #getTvShowRecommendations,
+              [
+                id,
+                page,
+              ],
+            ),
+            returnValue:
+                _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>.value(
+                    _FakeEither_0<_i8.Failure, List<_i16.TvShow>>(
+              this,
+              Invocation.method(
+                #getTvShowRecommendations,
+                [
+                  id,
+                  page,
+                ],
+              ),
+            )),
+          ) as _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>> searchTvShows(
+    String? query, [
+    int? page = 1,
+  ]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #searchTvShows,
+          [
+            query,
+            page,
+          ],
+        ),
+        returnValue:
+            _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>.value(
+                _FakeEither_0<_i8.Failure, List<_i16.TvShow>>(
+          this,
+          Invocation.method(
+            #searchTvShows,
+            [
+              query,
+              page,
+            ],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, String>> saveWatchlist(
+          _i17.TvShowDetail? tvShow) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #saveWatchlist,
+          [tvShow],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, String>>.value(
+            _FakeEither_0<_i8.Failure, String>(
+          this,
+          Invocation.method(
+            #saveWatchlist,
+            [tvShow],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, String>>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, String>> removeWatchlist(
+          _i17.TvShowDetail? tvShow) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #removeWatchlist,
+          [tvShow],
+        ),
+        returnValue: _i7.Future<_i2.Either<_i8.Failure, String>>.value(
+            _FakeEither_0<_i8.Failure, String>(
+          this,
+          Invocation.method(
+            #removeWatchlist,
+            [tvShow],
+          ),
+        )),
+      ) as _i7.Future<_i2.Either<_i8.Failure, String>>);
+  @override
+  _i7.Future<bool> isAddedToWatchlist(int? id) => (super.noSuchMethod(
+        Invocation.method(
+          #isAddedToWatchlist,
+          [id],
+        ),
+        returnValue: _i7.Future<bool>.value(false),
+      ) as _i7.Future<bool>);
+  @override
+  _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>
+      getWatchlistTvShows() => (super.noSuchMethod(
+            Invocation.method(
+              #getWatchlistTvShows,
+              [],
+            ),
+            returnValue:
+                _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>.value(
+                    _FakeEither_0<_i8.Failure, List<_i16.TvShow>>(
+              this,
+              Invocation.method(
+                #getWatchlistTvShows,
+                [],
+              ),
+            )),
+          ) as _i7.Future<_i2.Either<_i8.Failure, List<_i16.TvShow>>>);
+}
+
+/// A class which mocks [TvShowRemoteDataSource].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockTvShowRemoteDataSource extends _i1.Mock
+    implements _i18.TvShowRemoteDataSource {
+  MockTvShowRemoteDataSource() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i7.Future<List<_i19.TvShowModel>> getNowAiringTvShows([int? page = 1]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getNowAiringTvShows,
+          [page],
+        ),
+        returnValue:
+            _i7.Future<List<_i19.TvShowModel>>.value(<_i19.TvShowModel>[]),
+      ) as _i7.Future<List<_i19.TvShowModel>>);
+  @override
+  _i7.Future<List<_i19.TvShowModel>> getPopularTvShows([int? page = 1]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getPopularTvShows,
+          [page],
+        ),
+        returnValue:
+            _i7.Future<List<_i19.TvShowModel>>.value(<_i19.TvShowModel>[]),
+      ) as _i7.Future<List<_i19.TvShowModel>>);
+  @override
+  _i7.Future<List<_i19.TvShowModel>> getTopRatedTvShows([int? page = 1]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTopRatedTvShows,
+          [page],
+        ),
+        returnValue:
+            _i7.Future<List<_i19.TvShowModel>>.value(<_i19.TvShowModel>[]),
+      ) as _i7.Future<List<_i19.TvShowModel>>);
+  @override
+  _i7.Future<_i4.TvShowDetailModel> getTvShowDetail(int? id) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTvShowDetail,
+          [id],
+        ),
+        returnValue:
+            _i7.Future<_i4.TvShowDetailModel>.value(_FakeTvShowDetailModel_2(
+          this,
+          Invocation.method(
+            #getTvShowDetail,
+            [id],
+          ),
+        )),
+      ) as _i7.Future<_i4.TvShowDetailModel>);
+  @override
+  _i7.Future<List<_i19.TvShowModel>> getTvShowRecommendations(
+    int? id, [
+    int? page = 1,
+  ]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getTvShowRecommendations,
+          [
+            id,
+            page,
+          ],
+        ),
+        returnValue:
+            _i7.Future<List<_i19.TvShowModel>>.value(<_i19.TvShowModel>[]),
+      ) as _i7.Future<List<_i19.TvShowModel>>);
+  @override
+  _i7.Future<List<_i19.TvShowModel>> searchTvShows(
+    String? query, [
+    int? page = 1,
+  ]) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #searchTvShows,
+          [
+            query,
+            page,
+          ],
+        ),
+        returnValue:
+            _i7.Future<List<_i19.TvShowModel>>.value(<_i19.TvShowModel>[]),
+      ) as _i7.Future<List<_i19.TvShowModel>>);
+}
+
+/// A class which mocks [TvShowLocalDataSource].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockTvShowLocalDataSource extends _i1.Mock
+    implements _i20.TvShowLocalDataSource {
+  MockTvShowLocalDataSource() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i7.Future<String> insertWatchlist(_i21.TvShowTable? tvShow) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #insertWatchlist,
+          [tvShow],
+        ),
+        returnValue: _i7.Future<String>.value(''),
+      ) as _i7.Future<String>);
+  @override
+  _i7.Future<String> removeWatchlist(_i21.TvShowTable? tvShow) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #removeWatchlist,
+          [tvShow],
+        ),
+        returnValue: _i7.Future<String>.value(''),
+      ) as _i7.Future<String>);
+  @override
+  _i7.Future<_i21.TvShowTable?> getTvShowById(int? id) => (super.noSuchMethod(
+        Invocation.method(
+          #getTvShowById,
+          [id],
+        ),
+        returnValue: _i7.Future<_i21.TvShowTable?>.value(),
+      ) as _i7.Future<_i21.TvShowTable?>);
+  @override
+  _i7.Future<List<_i21.TvShowTable>> getWatchlistTvShows() =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getWatchlistTvShows,
+          [],
+        ),
+        returnValue:
+            _i7.Future<List<_i21.TvShowTable>>.value(<_i21.TvShowTable>[]),
+      ) as _i7.Future<List<_i21.TvShowTable>>);
 }
 
 /// A class which mocks [DatabaseHelper].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockDatabaseHelper extends _i1.Mock implements _i15.DatabaseHelper {
+class MockDatabaseHelper extends _i1.Mock implements _i22.DatabaseHelper {
   MockDatabaseHelper() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i7.Future<_i16.Database?> get database =>
-      (super.noSuchMethod(Invocation.getter(#database),
-              returnValue: Future<_i16.Database?>.value())
-          as _i7.Future<_i16.Database?>);
+  _i7.Future<_i23.Database?> get database => (super.noSuchMethod(
+        Invocation.getter(#database),
+        returnValue: _i7.Future<_i23.Database?>.value(),
+      ) as _i7.Future<_i23.Database?>);
   @override
-  _i7.Future<int> insertWatchlist(_i14.MovieTable? movie) =>
-      (super.noSuchMethod(Invocation.method(#insertWatchlist, [movie]),
-          returnValue: Future<int>.value(0)) as _i7.Future<int>);
+  _i7.Future<int> insertWatchlist(_i24.WatchlistTable? watchlist) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #insertWatchlist,
+          [watchlist],
+        ),
+        returnValue: _i7.Future<int>.value(0),
+      ) as _i7.Future<int>);
   @override
-  _i7.Future<int> removeWatchlist(_i14.MovieTable? movie) =>
-      (super.noSuchMethod(Invocation.method(#removeWatchlist, [movie]),
-          returnValue: Future<int>.value(0)) as _i7.Future<int>);
+  _i7.Future<int> removeWatchlist(_i24.WatchlistTable? watchlist) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #removeWatchlist,
+          [watchlist],
+        ),
+        returnValue: _i7.Future<int>.value(0),
+      ) as _i7.Future<int>);
   @override
-  _i7.Future<Map<String, dynamic>?> getMovieById(int? id) =>
-      (super.noSuchMethod(Invocation.method(#getMovieById, [id]),
-              returnValue: Future<Map<String, dynamic>?>.value())
-          as _i7.Future<Map<String, dynamic>?>);
+  _i7.Future<Map<String, dynamic>?> getWatchlistById(
+    int? id,
+    String? type,
+  ) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getWatchlistById,
+          [
+            id,
+            type,
+          ],
+        ),
+        returnValue: _i7.Future<Map<String, dynamic>?>.value(),
+      ) as _i7.Future<Map<String, dynamic>?>);
   @override
-  _i7.Future<List<Map<String, dynamic>>> getWatchlistMovies() =>
-      (super.noSuchMethod(Invocation.method(#getWatchlistMovies, []),
-              returnValue: Future<List<Map<String, dynamic>>>.value(
-                  <Map<String, dynamic>>[]))
-          as _i7.Future<List<Map<String, dynamic>>>);
+  _i7.Future<List<Map<String, dynamic>>> getWatchlist(String? type) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getWatchlist,
+          [type],
+        ),
+        returnValue: _i7.Future<List<Map<String, dynamic>>>.value(
+            <Map<String, dynamic>>[]),
+      ) as _i7.Future<List<Map<String, dynamic>>>);
 }
 
 /// A class which mocks [Client].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockHttpClient extends _i1.Mock implements _i17.Client {
+class MockHttpClient extends _i1.Mock implements _i5.Client {
   MockHttpClient() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i7.Future<_i4.Response> head(Uri? url, {Map<String, String>? headers}) =>
-      (super.noSuchMethod(Invocation.method(#head, [url], {#headers: headers}),
-              returnValue: Future<_i4.Response>.value(_FakeResponse()))
-          as _i7.Future<_i4.Response>);
-  @override
-  _i7.Future<_i4.Response> get(Uri? url, {Map<String, String>? headers}) =>
-      (super.noSuchMethod(Invocation.method(#get, [url], {#headers: headers}),
-              returnValue: Future<_i4.Response>.value(_FakeResponse()))
-          as _i7.Future<_i4.Response>);
-  @override
-  _i7.Future<_i4.Response> post(Uri? url,
-          {Map<String, String>? headers,
-          Object? body,
-          _i18.Encoding? encoding}) =>
+  _i7.Future<_i5.Response> head(
+    Uri? url, {
+    Map<String, String>? headers,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#post, [url],
-                  {#headers: headers, #body: body, #encoding: encoding}),
-              returnValue: Future<_i4.Response>.value(_FakeResponse()))
-          as _i7.Future<_i4.Response>);
+        Invocation.method(
+          #head,
+          [url],
+          {#headers: headers},
+        ),
+        returnValue: _i7.Future<_i5.Response>.value(_FakeResponse_3(
+          this,
+          Invocation.method(
+            #head,
+            [url],
+            {#headers: headers},
+          ),
+        )),
+      ) as _i7.Future<_i5.Response>);
   @override
-  _i7.Future<_i4.Response> put(Uri? url,
-          {Map<String, String>? headers,
-          Object? body,
-          _i18.Encoding? encoding}) =>
+  _i7.Future<_i5.Response> get(
+    Uri? url, {
+    Map<String, String>? headers,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#put, [url],
-                  {#headers: headers, #body: body, #encoding: encoding}),
-              returnValue: Future<_i4.Response>.value(_FakeResponse()))
-          as _i7.Future<_i4.Response>);
+        Invocation.method(
+          #get,
+          [url],
+          {#headers: headers},
+        ),
+        returnValue: _i7.Future<_i5.Response>.value(_FakeResponse_3(
+          this,
+          Invocation.method(
+            #get,
+            [url],
+            {#headers: headers},
+          ),
+        )),
+      ) as _i7.Future<_i5.Response>);
   @override
-  _i7.Future<_i4.Response> patch(Uri? url,
-          {Map<String, String>? headers,
-          Object? body,
-          _i18.Encoding? encoding}) =>
+  _i7.Future<_i5.Response> post(
+    Uri? url, {
+    Map<String, String>? headers,
+    Object? body,
+    _i25.Encoding? encoding,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#patch, [url],
-                  {#headers: headers, #body: body, #encoding: encoding}),
-              returnValue: Future<_i4.Response>.value(_FakeResponse()))
-          as _i7.Future<_i4.Response>);
+        Invocation.method(
+          #post,
+          [url],
+          {
+            #headers: headers,
+            #body: body,
+            #encoding: encoding,
+          },
+        ),
+        returnValue: _i7.Future<_i5.Response>.value(_FakeResponse_3(
+          this,
+          Invocation.method(
+            #post,
+            [url],
+            {
+              #headers: headers,
+              #body: body,
+              #encoding: encoding,
+            },
+          ),
+        )),
+      ) as _i7.Future<_i5.Response>);
   @override
-  _i7.Future<_i4.Response> delete(Uri? url,
-          {Map<String, String>? headers,
-          Object? body,
-          _i18.Encoding? encoding}) =>
+  _i7.Future<_i5.Response> put(
+    Uri? url, {
+    Map<String, String>? headers,
+    Object? body,
+    _i25.Encoding? encoding,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#delete, [url],
-                  {#headers: headers, #body: body, #encoding: encoding}),
-              returnValue: Future<_i4.Response>.value(_FakeResponse()))
-          as _i7.Future<_i4.Response>);
+        Invocation.method(
+          #put,
+          [url],
+          {
+            #headers: headers,
+            #body: body,
+            #encoding: encoding,
+          },
+        ),
+        returnValue: _i7.Future<_i5.Response>.value(_FakeResponse_3(
+          this,
+          Invocation.method(
+            #put,
+            [url],
+            {
+              #headers: headers,
+              #body: body,
+              #encoding: encoding,
+            },
+          ),
+        )),
+      ) as _i7.Future<_i5.Response>);
   @override
-  _i7.Future<String> read(Uri? url, {Map<String, String>? headers}) =>
-      (super.noSuchMethod(Invocation.method(#read, [url], {#headers: headers}),
-          returnValue: Future<String>.value('')) as _i7.Future<String>);
-  @override
-  _i7.Future<_i19.Uint8List> readBytes(Uri? url,
-          {Map<String, String>? headers}) =>
+  _i7.Future<_i5.Response> patch(
+    Uri? url, {
+    Map<String, String>? headers,
+    Object? body,
+    _i25.Encoding? encoding,
+  }) =>
       (super.noSuchMethod(
-              Invocation.method(#readBytes, [url], {#headers: headers}),
-              returnValue: Future<_i19.Uint8List>.value(_i19.Uint8List(0)))
-          as _i7.Future<_i19.Uint8List>);
+        Invocation.method(
+          #patch,
+          [url],
+          {
+            #headers: headers,
+            #body: body,
+            #encoding: encoding,
+          },
+        ),
+        returnValue: _i7.Future<_i5.Response>.value(_FakeResponse_3(
+          this,
+          Invocation.method(
+            #patch,
+            [url],
+            {
+              #headers: headers,
+              #body: body,
+              #encoding: encoding,
+            },
+          ),
+        )),
+      ) as _i7.Future<_i5.Response>);
   @override
-  _i7.Future<_i5.StreamedResponse> send(_i20.BaseRequest? request) =>
-      (super.noSuchMethod(Invocation.method(#send, [request]),
-              returnValue:
-                  Future<_i5.StreamedResponse>.value(_FakeStreamedResponse()))
-          as _i7.Future<_i5.StreamedResponse>);
+  _i7.Future<_i5.Response> delete(
+    Uri? url, {
+    Map<String, String>? headers,
+    Object? body,
+    _i25.Encoding? encoding,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #delete,
+          [url],
+          {
+            #headers: headers,
+            #body: body,
+            #encoding: encoding,
+          },
+        ),
+        returnValue: _i7.Future<_i5.Response>.value(_FakeResponse_3(
+          this,
+          Invocation.method(
+            #delete,
+            [url],
+            {
+              #headers: headers,
+              #body: body,
+              #encoding: encoding,
+            },
+          ),
+        )),
+      ) as _i7.Future<_i5.Response>);
   @override
-  void close() => super.noSuchMethod(Invocation.method(#close, []),
-      returnValueForMissingStub: null);
+  _i7.Future<String> read(
+    Uri? url, {
+    Map<String, String>? headers,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #read,
+          [url],
+          {#headers: headers},
+        ),
+        returnValue: _i7.Future<String>.value(''),
+      ) as _i7.Future<String>);
+  @override
+  _i7.Future<_i26.Uint8List> readBytes(
+    Uri? url, {
+    Map<String, String>? headers,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #readBytes,
+          [url],
+          {#headers: headers},
+        ),
+        returnValue: _i7.Future<_i26.Uint8List>.value(_i26.Uint8List(0)),
+      ) as _i7.Future<_i26.Uint8List>);
+  @override
+  _i7.Future<_i5.StreamedResponse> send(_i5.BaseRequest? request) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #send,
+          [request],
+        ),
+        returnValue:
+            _i7.Future<_i5.StreamedResponse>.value(_FakeStreamedResponse_4(
+          this,
+          Invocation.method(
+            #send,
+            [request],
+          ),
+        )),
+      ) as _i7.Future<_i5.StreamedResponse>);
+  @override
+  void close() => super.noSuchMethod(
+        Invocation.method(
+          #close,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
 }
